GetAll

+++ Models/Models/Post.cs
@@ -4,7 +4,7 @@ namespace Models
 {
     public abstract class Post
     {
-        private int _amount;
+        private double _amount;
         private double _price;
         private bool _isComplete;
         private IEnumerable<TransactionLine> _transactions;
@@ -12,7 +12,7 @@ namespace Models
         public Post()
         {
         }
-        public Post(int amount, double price, IEnumerable<TransactionLine> transactions)
+        public Post(double amount, double price, IEnumerable<TransactionLine> transactions)
         {
             _amount = amount;
             _price = price;
@@ -20,7 +20,7 @@ namespace Models
             _transactions = transactions;
         }
 
-        public int Amount { get { return _amount; } init { _amount = value; } }
+        public double Amount { get { return _amount; } init { _amount = value; } }
         public double Price { get { return _price; } init { _price = value; } }
         public bool IsComplete { get { return _isComplete; } set { _isComplete = value; } }
         public IEnumerable<TransactionLine> Transactions { get { return _transactions; } }
